+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/cti/catalog/processor/IntegrationProcessor.java
@@ -31,6 +31,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
 import com.wazuh.contentmanager.cti.catalog.utils.CategoryFormatter;
+import com.wazuh.contentmanager.utils.Constants;
 import com.wazuh.securityanalytics.action.WIndexIntegrationAction;
 import com.wazuh.securityanalytics.action.WIndexIntegrationRequest;
 import com.wazuh.securityanalytics.action.WIndexIntegrationResponse;
@@ -115,15 +116,16 @@ public class IntegrationProcessor extends AbstractProcessor {
             return;
         }
 
-        String id = doc.get("id").getAsString();
-        String name = doc.has("title") ? doc.get("title").getAsString() : "";
-        String description = doc.has("description") ? doc.get("description").getAsString() : "";
+        String id = doc.get(Constants.KEY_ID).getAsString();
+        String name = doc.has(Constants.KEY_TITLE) ? doc.get(Constants.KEY_TITLE).getAsString() : "";
+        String description =
+                doc.has(Constants.KEY_DESCRIPTION) ? doc.get(Constants.KEY_DESCRIPTION).getAsString() : "";
         String category = CategoryFormatter.format(doc, false);
 
         // Extract related rules to create the Threat Detector later.
         List<String> rules = new ArrayList<>();
-        if (doc.has("rules")) {
-            doc.get("rules").getAsJsonArray().forEach(item -> rules.add(item.getAsString()));
+        if (doc.has(Constants.KEY_RULES)) {
+            doc.get(Constants.KEY_RULES).getAsJsonArray().forEach(item -> rules.add(item.getAsString()));
         }
 
         try {